# internal/services/identity_service/docs/v1/swagger.yaml
definitions:
...
  CreateOrEditRoleDto:
    properties:
      grantedPermissions:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  CreateOrEditUserDto:
    properties:
      email:
        maxLength: 256
        type: string
      firstName:
        maxLength: 64
        minLength: 3
        type: string
      id:
        type: integer
      lastName:
        maxLength: 64
        minLength: 3
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      userName:
        maxLength: 256
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - userName
    type: object
  CreateRoleDto:
    properties:
      grantedPermissions:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  CreateUserDto:
    properties:
      email:
        maxLength: 256
        type: string
      firstName:
        maxLength: 64
        minLength: 3
        type: string
      id:
        type: integer
      lastName:
        maxLength: 64
        minLength: 3
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      userName:
        maxLength: 256
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - userName
    type: object
  EditRoleDto:
    properties:
      grantedPermissions:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  EditUserDto:
    properties:
      email:
        maxLength: 256
        type: string
      firstName:
        maxLength: 64
        minLength: 3
        type: string
      id:
        type: integer
      lastName:
        maxLength: 64
        minLength: 3
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      userName:
        maxLength: 256
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - userName
    type: object
...
  GetRoleByIdResult:
    properties:
      role:
        $ref: '#/definitions/CreateOrEditRoleDto'
    type: object
  GetRolesResult:
    properties:
      items:
        items:
          $ref: '#/definitions/RoleDto'
        type: array
      totalCount:
        type: integer
    type: object
  GetUserByIdResult:
    properties:
      user:
        $ref: '#/definitions/CreateOrEditUserDto'
    type: object
  GetUsersResult:
    properties:
      items:
        items:
          $ref: '#/definitions/UserDto'
        type: array
      totalCount:
        type: integer
    type: object
...
  RoleDto:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  UserDto:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      userName:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/identities/role:
    post:
      consumes:
      - application/json
      description: Create new role
      operationId: CreateRole
      parameters:
      - description: CreateRoleDto
        in: body
        name: CreateRoleDto
        schema:
          $ref: '#/definitions/CreateRoleDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RoleDto'
      security:
      - ApiKeyAuth: []
      summary: Create new role
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update role
      operationId: UpdateRole
      parameters:
      - description: EditRoleDto
        in: body
        name: EditRoleDto
        schema:
          $ref: '#/definitions/EditRoleDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RoleDto'
      security:
      - ApiKeyAuth: []
      summary: Update role
      tags:
      - Roles
  /api/v1/identities/role/{roleId}:
    delete:
      consumes:
      - application/json
      description: Delete role
      operationId: DeleteRole
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: Get role by id
      operationId: GetRoleById
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetRoleByIdResult'
      security:
      - ApiKeyAuth: []
      summary: Get role by id
      tags:
      - Roles
  /api/v1/identities/roles:
    get:
      consumes:
      - application/json
      description: Get all roles
      operationId: GetAllRoles
      parameters:
      - in: query
        name: filters
        type: string
      - in: query
        name: maxResultCount
        type: integer
      - in: query
        name: skipCount
        type: integer
      - in: query
        name: sorting
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetRolesResult'
      security:
      - ApiKeyAuth: []
      summary: Get all roles
      tags:
      - Roles
...
  /api/v1/identities/user:
    post:
      consumes:
      - application/json
      description: Create new user
      operationId: CreateUser
      parameters:
      - description: CreateUserDto
        in: body
        name: CreateUserDto
        schema:
          $ref: '#/definitions/CreateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserDto'
      security:
      - ApiKeyAuth: []
      summary: Create new user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user
      operationId: UpdateUser
      parameters:
      - description: EditUserDto
        in: body
        name: EditUserDto
        schema:
          $ref: '#/definitions/EditUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserDto'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
  /api/v1/identities/user/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete user
      operationId: DeleteUser
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user by id
      operationId: GetUserById
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetUserByIdResult'
      security:
      - ApiKeyAuth: []
      summary: Get user by id
      tags:
      - Users
...
  /api/v1/identities/users:
    get:
      consumes:
      - application/json
      description: Get all users
      operationId: GetAllUsers
      parameters:
      - in: query
        name: filters
        type: string
      - in: query
        name: maxResultCount
        type: integer
      - in: query
        name: skipCount
        type: integer
      - in: query
        name: sorting
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetUsersResult'
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"